+++ docs/app/Examples/elements/Button/Types/ButtonLabeledExample.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { Button, Icon } from 'stardust'
+import { Button, Icon, Label } from 'stardust'
 
 export default class ButtonLabeledExample extends Component {
   render() {
@@ -9,9 +9,9 @@ export default class ButtonLabeledExample extends Component {
           <Icon className='heart' /> Like
         </Button>
         {/* TODO: See issue #46 - button needs to render as a div */}
-        <a className='ui basic label'>
+        <Label basic link>
           3,000
-        </a>
+        </Label>
       </Button>
     )
   }