+++ docs/app/Examples/elements/Container/ContainerExamples.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import ContainerTypesExamples from './Types/ContainerTypesExamples'
+import ContainerVariationsExamples from './Variations/ContainerVariationsExamples'
+
+export default class ContainerExamples extends Component {
+  render() {
+    return (
+      <div>
+        <ContainerTypesExamples />
+        <ContainerVariationsExamples />
+      </div>
+    )
+  }
+}