+++ docs/app/Examples/elements/Header/States/HeaderDisabledExample.js
@@ -4,7 +4,7 @@ import { Header } from 'stardust'
 export default class HeaderDisabledExample extends Component {
   render() {
     return (
-      <Header.H2 className='disabled'>
+      <Header.H2 disabled>
         Disabled Header
       </Header.H2>
     )