+++ docs/app/Examples/elements/Header/Variations/HeaderDividingExample.js
@@ -4,7 +4,7 @@ import { Header } from 'stardust'
 export default class HeaderDividingExample extends Component {
   render() {
     return (
-      <Header.H3 className='dividing'>
+      <Header.H3 dividing>
         Dividing Header
       </Header.H3>
     )