+++ src/elements/Header/HeaderH2.js
@@ -1,19 +1,18 @@
-import React, { Component } from 'react'
+import React from 'react'
 import META from '../../utils/Meta'
 import _Header from './_Header'
 
-export default class HeaderH2 extends Component {
-
-  static _meta = {
-    library: META.library.semanticUI,
-    name: 'HeaderH2',
-    parent: 'Header',
-    type: META.type.element,
-  }
-
-  render() {
-    return (
-      <_Header {...this.props} _sdClass='sd-header-h2' _headerElement='h2' />
-    )
-  }
+function HeaderH2(props) {
+  return (
+    <_Header {...props} _sdClass='sd-header-h2' _headerElement='h2' />
+  )
 }
+
+HeaderH2._meta = {
+  library: META.library.semanticUI,
+  name: 'HeaderH2',
+  parent: 'Header',
+  type: META.type.element,
+}
+
+export default HeaderH2