+++ src/elements/Header/HeaderH5.js
@@ -1,18 +1,18 @@
-import React, { Component } from 'react'
+import React from 'react'
 import META from '../../utils/Meta'
 import _Header from './_Header'
 
-export default class HeaderH5 extends Component {
-  static _meta = {
-    library: META.library.semanticUI,
-    name: 'HeaderH5',
-    parent: 'Header',
-    type: META.type.element,
-  }
-
-  render() {
-    return (
-      <_Header {...this.props} _sdClass='sd-header-h5' _headerElement='h5' />
-    )
-  }
+function HeaderH5(props) {
+  return (
+    <_Header {...props} _sdClass='sd-header-h5' _headerElement='h5' />
+  )
 }
+
+HeaderH5._meta = {
+  library: META.library.semanticUI,
+  name: 'HeaderH5',
+  parent: 'Header',
+  type: META.type.element,
+}
+
+export default HeaderH5