+++ test/specs/elements/Label/Label-test.js
@@ -10,6 +10,7 @@ describe('Label Component', () => {
   common.hasUIClassName(Label)
   common.rendersChildren(Label)
 
+  common.propKeyOnlyToClassName(Label, 'basic')
   common.propKeyOnlyToClassName(Label, 'circular')
   common.propKeyOnlyToClassName(Label, 'floating')
   common.propKeyOnlyToClassName(Label, 'horizontal')
@@ -19,8 +20,8 @@ describe('Label Component', () => {
   common.propValueOnlyToClassName(Label, 'size')
 
   common.propKeyAndValueToClassName(Label, 'attached')
-  common.propKeyAndValueToClassName(Label, 'corner')
 
+  common.propKeyOrValueToClassName(Label, 'corner')
   common.propKeyOrValueToClassName(Label, 'pointing')
   common.propKeyOrValueToClassName(Label, 'ribbon')